refactor(FNP-60): move useAuth hook to hooks directory for better modularity
