+++ packages/client/src/hoc/AuthProvider/AuthProvider.tsx
@@ -1,22 +1,22 @@
 import type { ReactNode } from 'react';
-import { useContext, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import AuthAPI from 'src/services/api/auth-api';
 import type { Nullable } from 'src/shared/types/global';
 import type { TUser } from 'src/shared/types/user';
 import Helpers from 'src/utils/helpers';
 
-import type { AuthContextType } from './AuthProvider.context';
 import { AuthContext } from './AuthProvider.context';
 
 export const authAPI = new AuthAPI();
 
 export const AuthProvider = ({ children }: { children: ReactNode }) => {
   const [user, setUser] = useState<Nullable<TUser>>(null);
-  const [userIsLoading, setUserIsLoading] = useState(true);
+  const [userIsLoading, setUserIsLoading] = useState<boolean>(true);
 
-  const isAuthenticated = async () => {
-    setUserIsLoading(true);
+  useEffect(() => {
+    const isAuthenticated = async () => {
+      setUserIsLoading(true);
 
     try {
       const isAuth = localStorage.getItem('isAuth');
@@ -28,6 +28,10 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
     }
   };
 
+    isAuthenticated();
+  }, []);
+
+
   const updateUser = (user: TUser) => {
     localStorage.setItem('isAuth', 'true');
     setUser(user);
@@ -37,22 +41,11 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
     Helpers.Log('ERROR', err);
   };
 
-  useEffect(() => {
-    isAuthenticated();
-  }, []);
+
 
   return (
-    <AuthContext.Provider
-      value={{ user, userIsLoading, setUser, setUserIsLoading }}>
+    <AuthContext.Provider value={{ user, setUser, userIsLoading }}>
       {children}
     </AuthContext.Provider>
   );
-};
-
-export const useAuth = (): AuthContextType => {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
 };
