+++ packages/client/src/hooks/use-auth/index.ts
@@ -1,0 +1,1 @@
+export { useAuth } from 'src/hooks/use-auth/use-auth';
